Fix regressions. Add turns as Parameter 5. Reorganize GState to SarsaState.
The last commit included serveral build breaking regressions. Those where fixed by this commit.
This commit adds the possibility to submit the turns-count as Parameter 5. The default (if no parameter 5 is passed) is "20" rounds.
The GState and everything related is renamed to "SarsaState" to show that it is used in SarsaLambda. Furthormore SarsaState is now in the SarsaLambda-Package.